+++ gitoxide-core/src/index/entries.rs
@@ -1,4 +1,35 @@
+use crate::index::{parse_file, Options};
 use git_repository as git;
+use std::path::Path;
+
+pub fn entries(
+    index_path: impl AsRef<Path>,
+    mut out: impl std::io::Write,
+    Options { object_hash, format }: Options,
+) -> anyhow::Result<()> {
+    use crate::OutputFormat::*;
+    let file = parse_file(index_path, object_hash)?;
+
+    #[cfg(feature = "serde1")]
+    if let Json = format {
+        out.write_all(b"[\n")?;
+    }
+
+    let mut entries = file.entries().iter().peekable();
+    while let Some(entry) = entries.next() {
+        match format {
+            Human => to_human(&mut out, &file, entry)?,
+            #[cfg(feature = "serde1")]
+            Json => to_json(&mut out, &file, entry, entries.peek().is_none())?,
+        }
+    }
+
+    #[cfg(feature = "serde1")]
+    if let Json = format {
+        out.write_all(b"]\n")?;
+    }
+    Ok(())
+}
 
 #[cfg(feature = "serde1")]
 pub(crate) fn to_json(
